requirements.txt
pyModeS
geographiclib
spidev
adafruit-circuitpython-neopixel
RPi.GPIO
gpsd-py3
